sudo make coffee
Make sure to pip install the twitter library before trying this code. The instructions I chose to tell your program when to launch are yours to play with, I did not want to write a robotic message such as "Make coffee" because It´d be weird for my followers on Twitter... but you are free to pick the line you prefer. I´ll add more Twitter handling exceptions later...